[UPDATE] components hook refactor
