--- src/components/atoms/icon-button/style.css
@@ -1,15 +1,0 @@
-span.ripple {
-  background-color: rgba(0, 0, 0, 0.3);
-  border-radius: 50%;
-  position: absolute;
-  transform: scale(0);
-  animation: ripple .5s linear forwards;
-  z-index:1;
-}
-
-@keyframes ripple {
-  to {
-    transform: scale(1);
-    opacity: 0;
-  }
-}
