+++ src/components/atoms/icon/Icon.tsx
@@ -2,8 +2,10 @@ import { cn } from '@/lib/utils';
 import { DynamicIcon } from 'lucide-react/dynamic';
 import type { FC } from 'react';
 import type { IconProps } from './types';
+import { useIcon } from './useIcon';
 
-const Icon: FC<IconProps> = ({ name, color = 'text-accent', colorDark = 'text-accent', size = 24, className }) => {
+const Icon: FC<IconProps> = ({ ...props }) => {
+  const { name, color, colorDark, className, size } = useIcon(props);
   return <DynamicIcon name={name} className={cn(color, colorDark, className)} size={size} />;
 };
