+++ src/components/atoms/input/Input.tsx
@@ -1,147 +1,52 @@
-import { useRipple } from '@/components/utils/hooks/useRipple';
 import { cn } from '@/lib/utils';
 import type { VariantProps } from 'class-variance-authority';
-import { useEffect, useRef, useState } from 'react';
-import type { ChangeEvent, ComponentProps, FocusEvent, KeyboardEvent, RefObject } from 'react';
+import type { ComponentProps, FC, RefObject } from 'react';
+import Icon from '../icon';
 import { type InputProps, inputVariants, labelVariants } from './types';
-import './style.css';
-import Icon from '@atoms/icon';
+import { useInput } from './useInput';
+import '@/components/utils/styles/index.css';
 
-const formatAriaIds = (ids?: string | string[]) => (Array.isArray(ids) ? ids.join(' ') : ids);
-
-const IncrementDecrementButtons = ({
-  ref,
-  setHasValue
-}: { ref: RefObject<HTMLInputElement>; setHasValue: (value: boolean) => void }) => (
-  <div className='absolute right-2 top-0 bottom-0 flex flex-col justify-center items-center gap-[2px] z-10'>
-    <button
-      role='button'
-      aria-label='Increase value'
-      className='bg-gray-light-500 hover:bg-gray-light-700 dark:bg-gray-dark-600 dark:hover:bg-gray-dark-400 rounded-t-sm px-1 cursor-pointer outline-offset-0 dark:focus-visible:outline-white focus-visible:outline-secondary focus-visible:outline-1'
-      onClick={() => {
-        if (ref.current) {
-          const currentValue = parseFloat(ref.current.value !== '' ? ref.current.value : '0');
-          ref.current.value = (currentValue + 1).toString();
-          setHasValue(!!ref.current.value);
-        }
-      }}
-    >
-      <Icon name='chevron-up' color='text-text-light' colorDark='dark:text-text-dark' size={16} />
-    </button>
-    <button
-      role='button'
-      aria-label='Decrease value'
-      className='bg-gray-light-500 hover:bg-gray-light-700 dark:bg-gray-dark-600 dark:hover:bg-gray-dark-400 rounded-b-sm px-1 cursor-pointer outline-offset-0 dark:focus-visible:outline-white focus-visible:outline-secondary focus-visible:outline-1'
-      onClick={() => {
-        if (ref.current) {
-          const currentValue = parseFloat(ref.current.value !== '' ? ref.current.value : '0');
-          ref.current.value = (currentValue - 1).toString();
-          setHasValue(!!ref.current.value);
-        }
-      }}
-    >
-      <Icon name='chevron-down' color='text-text-light' colorDark='dark:text-text-dark' size={16} />
-    </button>
-  </div>
-);
-
-const PasswordToggleButton = ({
-  showPassword,
-  setShowPassword
-}: { showPassword: boolean; setShowPassword: (value: boolean) => void }) => (
-  <div className='absolute right-2 top-0 bottom-0 flex flex-col justify-center items-center gap-[2px] z-10'>
-    <button
-      role='button'
-      aria-label={!showPassword ? 'Show password' : 'Hide password'}
-      className='cursor-pointer outline-offset-0 dark:focus-visible:outline-white focus-visible:outline-secondary focus-visible:outline-1'
-      onClick={() => setShowPassword(!showPassword)}
-    >
-      <Icon
-        name={!showPassword ? 'eye' : 'eye-off'}
-        color='text-text-light'
-        colorDark='dark:text-text-dark'
-        className='hover:opacity-70'
-        size={20}
-      />
-    </button>
-  </div>
-);
-
-const Input = ({
-  rounded = false,
-  size = 'md',
-  label,
-  placeholder,
-  id,
-  type = 'text',
-  variant = 'regular',
-  className,
-  startContent,
-  endContent,
-  isFullWidth = false,
-  isRequired = false,
-  hint,
-  onFocus,
-  onBlur,
-  onKeyDown,
-  onChange,
-  value,
-  defaultValue,
-  disabled = false,
-  ariaDescribedBy,
-  ariaLabelledBy,
-  ...props
-}: Omit<VariantProps<typeof inputVariants>, 'state' | 'focused'> &
-  InputProps &
-  Omit<ComponentProps<'input'>, 'size' | 'id'> & {
-    ariaDescribedBy?: string | string[];
-    ariaLabelledBy?: string | string[];
-  }) => {
-  const [isFocused, setIsFocused] = useState(false);
-  const [hasValue, setHasValue] = useState(false);
-  const [showPassword, setShowPassword] = useState(false);
-  const containerRef = useRef<HTMLDivElement>(null);
-  const ref = useRef<HTMLInputElement>(null);
-
-  useRipple(containerRef);
-
-  useEffect(() => {
-    if (ref.current) {
-      ref.current.type = showPassword ? 'text' : 'password';
+const Input: FC<
+  Omit<VariantProps<typeof inputVariants>, 'state' | 'focused'> &
+    InputProps &
+    Omit<ComponentProps<'input'>, 'size' | 'id'> & {
+      ariaDescribedBy?: string | string[];
+      ariaLabelledBy?: string | string[];
     }
-  }, [showPassword]);
-
-  const handleFocus = (e: FocusEvent<HTMLInputElement>) => {
-    setIsFocused(true);
-    onFocus?.(e);
-  };
-
-  const handleBlur = (e: FocusEvent<HTMLInputElement>) => {
-    setIsFocused(false);
-    setHasValue(!!e.target.value);
-    onBlur?.(e);
-  };
-
-  const handleKeyDown = (e: KeyboardEvent<HTMLInputElement>) => {
-    if (['number', 'tel'].includes(type)) {
-      const allowedKeys = ['Backspace', 'Tab', 'ArrowLeft', 'ArrowRight', 'Delete', 'Enter', 'Home', 'End', '.'];
-      const isNumberKey = /^[0-9]$/.test(e.key);
-
-      if (!allowedKeys.includes(e.key) && !isNumberKey) {
-        e.preventDefault();
-      }
-      if (e.key === '.' && e.currentTarget.value.includes('.')) {
-        e.preventDefault();
-      }
-    }
-    onKeyDown?.(e);
-  };
-
-  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
-    setHasValue(!!e.target.value);
-    onChange?.(e, e.target.value);
-  };
-
+> = ({ ...props }) => {
+  const {
+    containerRef,
+    size,
+    rounded,
+    variant,
+    isFocused,
+    type,
+    isFullWidth,
+    label,
+    hint,
+    disabled,
+    className,
+    ref,
+    id,
+    hasValue,
+    placeholder,
+    startContent,
+    endContent,
+    isRequired,
+    formatAriaIds,
+    ariaDescribedBy,
+    ariaLabelledBy,
+    handleBlur,
+    handleChange,
+    handleFocus,
+    handleKeyDown,
+    showPassword,
+    value,
+    defaultValue,
+    setHasValue,
+    setShowPassword,
+    ...rest
+  } = useInput(props);
   const getIconByHintType = (type: string) => {
     switch (type) {
       case 'error':
@@ -154,7 +59,63 @@ const Input = ({
         return <Icon name='info' color='text-gray-light-700' colorDark='dark:text-gray-dark-300' size={16} />;
     }
   };
+  const IncrementDecrementButtons = ({
+    ref,
+    setHasValue
+  }: { ref: RefObject<HTMLInputElement>; setHasValue: (value: boolean) => void }) => (
+    <div className='absolute right-2 top-0 bottom-0 flex flex-col justify-center items-center gap-[2px] z-10'>
+      <button
+        role='button'
+        aria-label='Increase value'
+        className='bg-gray-light-500 hover:bg-gray-light-700 dark:bg-gray-dark-600 dark:hover:bg-gray-dark-400 rounded-t-sm px-1 cursor-pointer outline-offset-0 dark:focus-visible:outline-white focus-visible:outline-secondary focus-visible:outline-1'
+        onClick={() => {
+          if (ref.current) {
+            const currentValue = parseFloat(ref.current.value !== '' ? ref.current.value : '0');
+            ref.current.value = (currentValue + 1).toString();
+            setHasValue(!!ref.current.value);
+          }
+        }}
+      >
+        <Icon name='chevron-up' color='text-text-light' colorDark='dark:text-text-dark' size={16} />
+      </button>
+      <button
+        role='button'
+        aria-label='Decrease value'
+        className='bg-gray-light-500 hover:bg-gray-light-700 dark:bg-gray-dark-600 dark:hover:bg-gray-dark-400 rounded-b-sm px-1 cursor-pointer outline-offset-0 dark:focus-visible:outline-white focus-visible:outline-secondary focus-visible:outline-1'
+        onClick={() => {
+          if (ref.current) {
+            const currentValue = parseFloat(ref.current.value !== '' ? ref.current.value : '0');
+            ref.current.value = (currentValue - 1).toString();
+            setHasValue(!!ref.current.value);
+          }
+        }}
+      >
+        <Icon name='chevron-down' color='text-text-light' colorDark='dark:text-text-dark' size={16} />
+      </button>
+    </div>
+  );
 
+  const PasswordToggleButton = ({
+    showPassword,
+    setShowPassword
+  }: { showPassword: boolean; setShowPassword: (value: boolean) => void }) => (
+    <div className='absolute right-2 top-0 bottom-0 flex flex-col justify-center items-center gap-[2px] z-10'>
+      <button
+        role='button'
+        aria-label={!showPassword ? 'Show password' : 'Hide password'}
+        className='cursor-pointer outline-offset-0 dark:focus-visible:outline-white focus-visible:outline-secondary focus-visible:outline-1'
+        onClick={() => setShowPassword(!showPassword)}
+      >
+        <Icon
+          name={!showPassword ? 'eye' : 'eye-off'}
+          color='text-text-light'
+          colorDark='dark:text-text-dark'
+          className='hover:opacity-70'
+          size={20}
+        />
+      </button>
+    </div>
+  );
   return (
     <div className={`flex flex-col gap-2${type === 'hidden' ? ' hidden' : ''}`}>
       <div
@@ -186,7 +147,12 @@ const Input = ({
           className
         )}
         onClick={() =>
-          document.activeElement !== ref.current && type !== 'number' && type !== 'password' && ref.current?.focus()
+          ref &&
+          'current' in ref &&
+          document.activeElement !== ref.current &&
+          type !== 'number' &&
+          type !== 'password' &&
+          ref.current?.focus()
         }
       >
         <label
@@ -219,7 +185,7 @@ const Input = ({
         <div className='flex w-full justify-between gap-4'>
           {startContent && startContent}
           <input
-            {...props}
+            {...rest}
             ref={ref}
             id={id}
             type={type}
